requirements.txt
matplotlib
scikit-learn
tensorflow-gpu==1.14.0